Refactor regexes and mappings class and constants.
