+++ cpp/CMakeLists.txt
@@ -272,6 +272,7 @@ set (
   "src/phonenumbers/phonenumber.pb.cc"   # Generated by Protocol Buffers.
   "src/phonenumbers/phonenumberutil.cc"
   "src/phonenumbers/regex_based_matcher.cc"
+  "src/phonenumbers/regexpsandmappings.cc"
   "src/phonenumbers/regexp_cache.cc"
   "src/phonenumbers/shortnumberinfo.cc"
   "src/phonenumbers/string_byte_sink.cc"
